% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{abort_changes_msg}
\alias{abort_changes_msg}
\title{Error message, for when menu has been exited
or changes have been aborted}
\usage{
abort_changes_msg()
}
\value{
cli messages:
\itemize{
\item changing metadata has been aborted
\item no changes made
}
}
\description{
Error message, for when menu has been exited
or changes have been aborted
}
\keyword{internal}
