% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_meta.R
\name{add_meta}
\alias{add_meta}
\title{Add a new variable to the metadata
The function:
\itemize{
\item creates a list-object with metadata
\item ensures that required fields are complete,
\item allows to add more fields to the list
\item tries to ensure proper formatting and valid entries
\item updates current metadata yaml
}}
\usage{
add_meta(
  var = NULL,
  oberthema = NULL,
  unterthema = NULL,
  var_typ = NULL,
  umfrage_item = NULL,
  itemformulierung = NULL,
  skalenberechnung = NULL,
  umfrage_filter = NULL,
  quelle = NULL,
  status = "aktiv",
  values = NULL,
  ...,
  yaml_path = system.file("qm_sekII_metadata.yaml", package = "metaSABSEB")
)
}
\arguments{
\item{var}{Character string. Defines variable name.
No empty spaces, snake case (always required)}

\item{oberthema}{Character string. Defines topic of survey question
(always required)}

\item{unterthema}{Character string. (optional)}

\item{var_typ}{Character string. Describes type of survey question
(Einzelitem, Item-Gruppe, Skala) (always required)}

\item{umfrage_item}{Character string. Defines survey question
(always required)}

\item{itemformulierung}{Character string. Further defines the survey
question. (required if var_typ = Item-Gruppe)}

\item{skalenberechnung}{Character string.
(only required if var_typ == Skala)}

\item{umfrage_filter}{Character string. Defines survey filters.
-> what is required for the item to be shown in the survey (optional)}

\item{quelle}{Character string. Source for construction of survey item
(optional)}

\item{status}{Character string. Defines whether variable is actively used
or deprecated. Default when adding a new variable to metadata is "aktiv"}

\item{values}{List-Object. Lists encoding of all values.
(required if values are encoded)}

\item{...}{Character string. Allows to add more metadata entries.
Must be a named argument.}

\item{yaml_path}{Character string specifying the path to the metadata YAML
file. By default, the package's metadata YAML file is used. This argument
can be changed, e.g. for testing with a temporary YAML file.}
}
\description{
Add a new variable to the metadata
The function:
\itemize{
\item creates a list-object with metadata
\item ensures that required fields are complete,
\item allows to add more fields to the list
\item tries to ensure proper formatting and valid entries
\item updates current metadata yaml
}
}
\examples{
\dontrun{
metaSABSEB::add_meta(
  var = "TESTVARIABLE",
  oberthema = "Lehrpersonen",
  var_typ = "Einzelitem: Nettigkeit",
  umfrage_item = "Wie nett finden Sie Ihre Lehrperson?",
  values = list(
    "1" = "gar nicht",
    "2" = "eher nicht",
    "3" = "weder noch",
    "4" = "eher",
    "5" = "sehr"
  )
)
}


}
