% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_meta.R
\name{change_meta}
\alias{change_meta}
\title{Change metadata for a variable that is already registered
The function:
\itemize{
\item reads the metadata yaml
\item locates the desired variable and field
\item shows the current metadata entry
\item asks for confirmation to change it
\item changes the metadata yaml on disk
}}
\usage{
change_meta(
  var = NULL,
  field = NULL,
  new_entry = NULL,
  yaml_path = system.file("qm_sekII_metadata.yaml", package = "metaSABSEB")
)
}
\arguments{
\item{var}{A character string specifying the variable name}

\item{field}{A character string specifying the metadata field
to be changed}

\item{new_entry}{String or numeric, new value for the metadata field}

\item{yaml_path}{Character string specifying the path to the metadata YAML
file. By default, the package's metadata YAML file is used. This argument
can be changed, e.g. for testing with a temporary YAML file.}
}
\description{
Change metadata for a variable that is already registered
The function:
\itemize{
\item reads the metadata yaml
\item locates the desired variable and field
\item shows the current metadata entry
\item asks for confirmation to change it
\item changes the metadata yaml on disk
}
}
\examples{
\dontrun{
metaSABSEB::change_meta(
  var = "TESTVARIABLE",
  field = "umfrage_item",
  new_entry = "Wie zufrieden sind Sie mit Ihrer Lehrperson?"
)
}

}
