% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{confirm_replacement}
\alias{confirm_replacement}
\title{Confirm that an existing entry should be replaced
Compares current metadata entry to new entry
Asks interactively, if the entry should be replaced
Replaces entry on disk or}
\usage{
confirm_replacement(yaml_path, meta, var, field, new_entry)
}
\arguments{
\item{yaml_path}{Path to metadata}

\item{meta}{A named list as returned by \code{\link[yaml:read_yaml]{yaml::read_yaml()}}.}

\item{var}{A character string specifying the variable that should be changed}

\item{field}{A character string specyifing the field
that should be changed for the variable defined in \code{var}}

\item{new_entry}{String or numeric:
Metadata entry that should be used to replace old entry.}
}
\value{
description
}
\description{
Confirm that an existing entry should be replaced
Compares current metadata entry to new entry
Asks interactively, if the entry should be replaced
Replaces entry on disk or
}
