% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_meta.R
\name{delete_meta}
\alias{delete_meta}
\title{This function is used to delete metadata entries
May be used if an entry was wrongfully added}
\usage{
delete_meta(
  var = NULL,
  yaml_path = system.file("qm_sekII_metadata.yaml", package = "metaSABSEB")
)
}
\arguments{
\item{var}{Character string. Defines variable for which
the metadata should be deleted}

\item{yaml_path}{Character string specifying the path to the metadata YAML
file. By default, the package's metadata YAML file is used. This argument
can be changed, e.g. for testing with a temporary YAML file.}
}
\description{
This function is used to delete metadata entries
May be used if an entry was wrongfully added
}
\examples{
\dontrun{metaSABSEB::delete_meta("TESTVAR")}

}
