% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_meta.R
\name{get_meta}
\alias{get_meta}
\title{Retrieve metadata fields for a specified variable.
If no field is specified, all metadata is returned.}
\usage{
get_meta(
  var,
  field = NULL,
  refresh = FALSE,
  yaml_path = system.file("qm_sekII_metadata.yaml", package = "metaSABSEB")
)
}
\arguments{
\item{var}{A character string specifying the variable name}

\item{field}{A character string specifying the field to retrieve}

\item{refresh}{Logical. If TRUE,
the metadata cache is reloaded from disk.
Use this if the underlying metadata file has changed
(e.g. during development or after updating the package).
Default is FALSE.}

\item{yaml_path}{Character string specifying the path to the metadata YAML
file. By default, the package's metadata YAML file is used. This argument
can be changed, e.g. for testing with a temporary YAML file. description}
}
\value{
If \code{field = NULL}, a named list containing all metadata
associated with \code{var}. Otherwise the value of the requested
metadata field
}
\description{
Retrieve metadata fields for a specified variable.
If no field is specified, all metadata is returned.
}
\examples{
get_meta(var = "EFZ_Typ")
get_meta(var = "Anst_best", "quelle")

}
